requirements.txt
SQLAlchemy>=1.4
streamlit>=1.42.2
python-dotenv==1.0.1
langchain-openai>=0.3.7
langchain-community>=0.3.18
langchain>=0.3.19
langfuse>=2.59.7
langgraph>=0.3.2
faiss-cpu>=1.10.0
duckduckgo-search==8.1.1
yfinance>=0.2.61
sentence-transformers>=2.2.2
torch>=1.9.0
transformers>=4.21.0
plotly>=5.0.0
